// src/entities/survey/enum.ts
export enum SurveyLevelLabel {
  beginner = 'Начинающие',
  intermediate = 'Средний уровень',
  advanced = 'Продвинутые',
  any = 'Любой уровень',
}

export enum SurveyWeekdayLabel {
  mon = 'Пн',
  tue = 'Вт',
  wed = 'Ср',
  thu = 'Чт',
  fri = 'Пт',
  sat = 'Сб',
  sun = 'Вс',
}